% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DensityFunctions.R, R/DensityFunctions2.R
\name{dist_function}
\alias{dist_function}
\title{Distance function}
\usage{
dist_function(start_L, start_K, x0, intercept, range)

dist_function(start_L, start_K, x0, intercept, range)
}
\arguments{
\item{start_L}{output parameter from the logistic model function}

\item{start_K}{output parameter from the logistic model function}

\item{intercept}{denisty at x = 0 predicted from the logistic model function}

\item{range}{range over which to calculate the density}

\item{xo}{output parameter from the logistic model function}

\item{start_L}{output parameter from the logistic model function}

\item{start_K}{output parameter from the logistic model function}

\item{xo}{output parameter from the logistic model function}

\item{intercept}{denisty at x = 0 predicted from the logistic model function}

\item{range}{range over which to calculate the density}
}
\description{
This function calculates the absolute density from a frame by correcting for
the sighting function for the species.

This function calculates the absolute density from a frame by correcting for
the sighting function for the species.
}
\examples{
dist_function()
dist_function()
}
\keyword{logistic}
\keyword{model}
