% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DensityFunctions.R, R/DensityFunctions2.R
\name{frame_density}
\alias{frame_density}
\title{Absolute density for a frame}
\usage{
frame_density(start_L, start_K, x0, intercept, range_bin, density,
  near_limit = 1)

frame_density(start_L, start_K, x0, intercept, range_bin, density,
  near_limit = 1)
}
\arguments{
\item{start_L}{output parameter from the logistic model function}

\item{start_K}{output parameter from the logistic model function}

\item{range_bin}{series of ranges of densities from a frame;}

\item{density}{observed densities by range in a frame}

\item{xo}{output parameter from the logistic model function}

\item{start_L}{output parameter from the logistic model function}

\item{start_K}{output parameter from the logistic model function}

\item{xo}{output parameter from the logistic model function}

\item{range_bin}{series of ranges of densities from a frame;}

\item{density}{observed densities by range in a frame}
}
\description{
This function calculates the absolute density from a frame by correcting for
the sighting function for the species.

This function calculates the absolute density from a frame by correcting for
the sighting function for the species.
}
\examples{
frame_density()
frame_density()
}
\keyword{logistic}
\keyword{model}
