% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DensityFunctions.R, R/DensityFunctions2.R
\name{logistic_fit}
\alias{logistic_fit}
\title{A logistic fitting function}
\usage{
logistic_fit(range, par)

logistic_fit(range, par)
}
\arguments{
\item{start_L}{Maximum length}

\item{x0}{initial size}

\item{start_K}{carrying capacity}

\item{start_L}{Maximum length}

\item{x0}{initial size}

\item{start_K}{carrying capacity}
}
\description{
This function fits the logistic function.

This function fits the logistic function.
}
\examples{
logistic_fit()
logistic_fit()
}
\keyword{logisitic}
\keyword{model}
