% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DensityFunctions.R, R/DensityFunctions2.R
\name{zero_fill}
\alias{zero_fill}
\title{Zero filled data}
\usage{
zero_fill(d_data, species)

zero_fill(d_data, species)
}
\arguments{
\item{d_data}{set of density data that includes all unique combinations}

\item{species}{species of interest}

\item{d_data}{set of density data that includes all unique combinations}

\item{species}{species of interest}
}
\description{
This function takes the unique deployment, frame, range combinations and
attaches the density estimates for a species where they occur and fills
the rest of the unique deployment, frames and range combinations with
zeros..

This function takes the unique deployment, frame, range combinations and 
attaches the density estimates for a species where they occur and fills
the rest of the unique deployment, frames and range combinations with
zeros..
}
\examples{
frame_density()
frame_density()
}
\keyword{logistic}
\keyword{model}
