% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DensityFunctions.R
\name{logistic_modlike}
\alias{logistic_modlike}
\title{A negative log-likelihood function}
\usage{
logistic_modlike(par, range, ob_dens, weights = 1)
}
\arguments{
\item{start_L}{Maximum length}

\item{x0}{initial size}

\item{start_K}{carrying capacity}
}
\description{
This function estimates the likelihood of your parameters
}
\examples{
logistic_modlike()
}
\keyword{logistic}
\keyword{model}
